unified fz_cookie->progress and fz_cookie->progress_max to both be of size_t type: the progress should not overflow, just like its progress_max limit value.

+++ include/mupdf/fitz/device.h
@@ -293,7 +293,7 @@ enum
 typedef struct
 {
 	volatile int abort;
-	int progress;
+	size_t progress;
 	size_t progress_max; /* (size_t)-1 for unknown */
 	int errors;
 	int incomplete;
